Refactor the idp selection
